requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
transformers==4.48.2
torch --index-url https://download.pytorch.org/whl/cpu
pillow
requests
accelerate
